I fix the %lld in printf code, VOLT_DEBUG code has not been runed.

+++ src/ee/executors/indexcountexecutor.cpp
@@ -405,7 +405,7 @@ bool IndexCountExecutor::p_execute(const NValueArray &params)
         }
     }
     rkRes = rkEnd - rkStart - 1 + leftIncluded + rightIncluded;
-    VOLT_DEBUG("ANSWER %lld = %lld - %lld - 1 + %d + %d\n", rkRes, rkEnd, rkStart, leftIncluded, rightIncluded);
+    VOLT_DEBUG("ANSWER %ld = %ld - %ld - 1 + %d + %d\n", (long)rkRes, (long)rkEnd, (long)rkStart, leftIncluded, rightIncluded);
     tmptup.setNValue(0, ValueFactory::getBigIntValue( rkRes ));
     m_outputTable->insertTuple(tmptup);
